Fix async bug and update badge on back/forward
Relying on an asynchronous call to the chrome.storage API to add the
MutationSummary observer meant that the replacement might not run if the
page loaded before the callback executes. This has been fixed by moving
the mutation processing inside the storage callback, which does not
appear to have a noticeable performance penalty.